split up payment-methods.js file into smaller bits
- adds more inline docs
- improves/adds more type-defs
